KernelLtpStagingTest does not run stable tests
When KernelLtpStagingTest was first defined, it was used as a tot test and the stable version was pinned.
Now we have separated the conception of prod and test for pinned stable version, and additionally have vts-kernel running as tot tests, KernelLtpStagingTest should now only contains staging tests.

After this CL, KernelLtpStagingTest will only run tests that are in staging list.

Test: run vts -m KernelLtpStagingTest
Bug: 35882153
Change-Id: Ief6fd5dd84e3796ce41df2661eb3aa1bfa6ae12b