Additional work on procfs tests
For files that have clean printf lines identified in the kernel,
use the parse library which acts as the reverse of printf.

For some files like stat and vmalloc info, use the kernel functions
as a guide to refine parsing rules.

Used parse instead of ply for simple files.

Added simple existence check for all files.

Added new file tests.

Bug: 37153161
Bug: 35773187
Test: run tests
Change-Id: If92967511b04ec60f59a6fa4d790d4c0d698e902