LTP uses a stable list instead of a staging list
Previously, LTP uses a staging list to mark unstable tests.
The downside of doing so is assuming all tests that are not in disabled list are stable.
This is problematic when new test script or test cases are added.

After this CL, LTP runner will read a stable list of test names.
These tests have been tested on various device for multiple times and are known to be stable.

Change-Id: Id08f8b7bf33296e4be75111d07f86f45a1666e70
Fix: 62723431
Test: make vts; vts-tradefed; run vts-kernel
Signed-off-by: Yuexi Ma <yuexima@google.com>