remove disabled tests, pselect01_64 from stable list
Some tests are defined in the disabled list without bitness, but
included in the stable set with bitness. This causes the tests to be run
as part of the stable set. Remove the these tests from the stable set.

The pselect01_64 test is not reliable currently, remove from stable set.

Also remove mm.thp01_32bit from stable as it is defined in the disabled
list. It does not run in stable, but it is confusing nonetheless.

Bug: 66336091
Bug: 63254476
Test: run vts-kernel -m VtsKernelLtp
Merged-In: I91d4b5fc7ad6a4874219040fccad3c28fa1b85f2
Change-Id: I91d4b5fc7ad6a4874219040fccad3c28fa1b85f2