kselftest: provide separate arg for testsuite
The initializer for LinuxKselftestTestcase currently derives the
testsuite (directory location) of the test by looking for the first
directory level in the path. For test suites that are more than one
directory level deep this causes the test harness to not change
directory all the way into the directory with the test.

Modify the LinuxKselftestTestcase class to take a separate argument in
its initializer for the testsuite name. The second argument is modified
to be the the test command.

Bug: 31590178
Test: run vts-kernel -m VtsKernelLinuxKselftestStaging
Test: run vts-kernel -m VtsKernelLinuxKselftest
Change-Id: Ib92c7a837db2ae8ea102fb500cbf6b497da926c1