ltp: Run restorecon on LTPDIR to make sure its properly labelled
SELinux policy is updated to allow kernel to read files from ltp working
directory in order to do loopback filesystem mounts. This means, we need
to make sure that the LTPDIR is also appropriately labelled when it is
created everytime. Do that by running 'restorecon' on the directory
immediately after its creation.

Bug: 73220071
Test: run vts-kernel -m VtsKernelLtp -t syscalls.fchown04

Change-Id: I8a65067e6f8c62a92f8d06055a8640ef394f70c4
Signed-off-by: Sandeep Patil <sspatil@google.com>