Make qtaguid test static link to libcutils
The qtaguid native test is using libcutils api which may be different
depend on the VNDK version. in order to make the test compatible cross
devices and release, modify the Android.mk file to make it static link
to libcutils so it will not fail because of different VNDK api.

Bug:77678121
test: vts_test_binary_qtaguid_module show up in
      data/nativetest/vendor/ directory and test pass.
Change-Id: Ifccc0cdf1ba50a57714c3131dde051d61658511c