Directly test the qtaguid module with out libcutil
With the new changes to support bpf traffic accounting, the tag/untag
helper in libcutils are not directly linked to qtaguid module. It will
decide which module to use depend on the avalaible system on devices.
But we still want to test the qtaguid functionality to make sure it
works properly if the module is enabled. So reroute the test to directly
call qtaguid legacy API and all the test should be still functioning
properly with this change.

Test: run vts -m vtsKernelQtaguidTest
Bug: 30950746
Change-Id: I747804c93bee1666b039fe743728b72e11d78bc9