Remove the netd integration test from vts-kernel.
This test is not appropriate for VTS because it tests the
behaviour of netd, not the behaviour of kernel interfaces.
Thus, when upgrading a device, for example, from 8.1 to 9.0,
the 8.1 test's expectations are incorrect because the correct
behaviour changed when the OS was upgraded.

This test is being moved to cts-on-gsi.

While I'm at it, also skip the ResolverTest and TransportTest
test cases. These can be flaky, take a long time to run, and
don't directly exercise any kernel functionality.

Bug: 77680167
Test: make -j64 vts && vts-tradefed run commandAndExit CtsOnGsiKernelNetdTest.config
Change-Id: I78761835e950839a2d309fdd509b2493db0bec23