Test to check atomic bpf map swap
To safely drain the eBPF map from userspace without deleting the kernel
eBPF program we need to swap the kernel eBPF map that the kernel
program is using during run time. This operation can be done atomically
using membarrier syscall and this test verifies the behavior of the
syscall is consistent.

Bug: 79171384
Test: ./vts_test_binary_bpf_module/vts_test_binary_bpf_module
Change-Id: I798694ea5bd5e3f56ec9c9c53bf4588efab3c951