ltp: add mandatory flag to LTP test configuration
LTP tests often are written such that they are skipped if the device
configuration renders the test inapplicable. This may be permissible in
some cases but in many others LTP tests functionality that Android
depends upon, therefore these tests must pass, and a skip should be
considered a failure.

Add a flag in the stable list which can be used to indicate that a
result of TCONF shall indicate failure rather than skip.

Bug: 111764751
Change-Id: Idf230b6f4f21120bc59209603ec0d180bc45d510
Signed-off-by: Steve Muckle <smuckle@google.com>