Change bpf_native_test timeout mechanism
Current clock method to calculate time interval will
have tolerance, change to use system_clock,
it can calculate time interval more precisely.

This is because std::clock() measures cpu time not real
wall clock time, and thus may be slower if idle, or faster
(by a factor of up to the number of cores/hyperthreads on
the device) if busy and multi-threaded.

Bug: 138082942
Test: run vts -m VtsKernelNetBpfTest
  and 'atest bpf_module_test'
Change-Id: I1224a0508f3240f9b01ccaf7e648e711fefda60e