Remove VtsKernelRootDirTest
This test is obsolete.  It was originally added for ensuring that
ramdisks contained the appropriate contents for devices that do not
include rootdir in system.img, which is separately tested in
VtsSecurityAvbTest.  As of Android Q, the rootdir is never included in
a ramdisk and always built as a part of system.img, so this test is no
longer needed.

Bug: 140313207
Test: build
Change-Id: I4723a2f35533fc10aa7974bef7016db7dfc5e6a0