Add ext4 requirement for /metadata
[cherry-picked from aosp/1270615]

During the first stage init, it is possible that fsck is triggered for
the metadata device.  In that case, the first stage init assumes that
the device is formatted as ext4, but there is no test to enforce that.

This patch introduces a VTS test that ensures the metadata device to be
formatted as ext4.

Test: atest vts_virtual_ab_test
Bug: 148177065
Signed-off-by: Alessio Balsini <balsini@google.com>
Change-Id: I945173859365937d9397144bdbc8bd8ead44a8b2
Merged-In: I945173859365937d9397144bdbc8bd8ead44a8b2