VtsKernelEncryptionTest: add FBE randomness test
Test that if the device uses FBE and we create a file in one of the
device's encrypted directories, the ciphertext seems to be random.

This isn't as strong a test as the FBE correctness tests.  However, this
is still a really useful test because it applies regardless of the
encryption format and keys.  Thus it runs even on old devices, including
ones that used a vendor-specific encryption format.

Test: 'atest vts_kernel_encryption_test' on cuttlefish and coral
Bug: 151100202
Change-Id: I3fb463b7e54c6922744cb61a5e476ba993d38e21