VtsKernelEncryptionTest: add metadata encryption randomness test
Test that if the device uses metadata encryption, then the filesystem
superblock on the userdata partition appears to be encrypted.

This isn't as strong a test as the dm-default-key correctness tests.
However, this is still a useful test because it applies regardless of
the encryption format and key.  Thus it runs even on old devices,
including ones that used a vendor-specific encryption format.

Test: 'atest vts_kernel_encryption_test' on cuttlefish and coral
Bug: 151100202
Change-Id: I56a48d89638949491a8e9bddc46cfa4b9788a533