VtsKernelEncryptionTest: define GetFilesystemInfo() and move to utils.cpp
Move the code to get information about a filesystem (fs block device,
uuid, type, and raw block device) into a utility function, so that it's
easier to change the way that we get this information.

Also rename "raw partition" to "raw block device", in preparation of the
following change which recognizes that what we want is not necessarily
the physical partition.

Test: tested together with I69f1fd9920fc261822a43487ed51f7cc6ef8d74a
Bug: 154446419
Change-Id: I7ec55b5531aa3f271eb31521064027b2d1995a56
Merged-In: I7ec55b5531aa3f271eb31521064027b2d1995a56