VtsKernelEncryptionTest: fix getting the raw block device
The encryption tests fail when the system is running from a DSU (Dynamic
System Update), since they get the raw block device of "/data" from the
FstabEntry.  In the DSU case, the FstabEntry lists the block device as
"userdata_gsi".  This isn't a path to a block device, but rather it's
the name of the logical partition (a partition assembled from free space
on the real userdata filesystem using dm-linear).

But I think using the fstab's listed block device is wrong anyway
because when checkpointing is used on ext4, there is a stack:

    userdata filesystem
    -> dm-default-key (only when metadata encryption is enabled)
    -> dm-bow
    -> dm-linear (only when userdata is a logical partition)
    -> physical userdata partition

Since dm-default-key is closely tied to the filesystem, it's
intentionally placed above dm-bow.  Thus, sector 0 may contain the
dm-bow log rather than the encrypted filesystem sector 0 as expected by
the metadata encryption randomness test.

To solve both these problems, stop using the fstab to get the raw block
device.  Instead, get it by checking if the mounted block device is a
dm-default-key device, and if so "unwrapping" it.

This gives us the correct block device in the stack to use for
encryption testing.

Test: Tested on Cuttlefish:

    $ANDROID_HOST_OUT/bin/launch_cvd \
            -daemon -data_policy always_create -blank_data_image_mb 8192
    atest vts_kernel_encryption_test
    adb root
    adb shell gsi_tool install \
            --gsi-size $(stat -c %s $ANDROID_PRODUCT_OUT/system.img) \
            < $ANDROID_PRODUCT_OUT/system.img
    sleep 60
    atest vts_kernel_encryption_test

Bug: 154446419
Change-Id: I69f1fd9920fc261822a43487ed51f7cc6ef8d74a
Merged-In: I69f1fd9920fc261822a43487ed51f7cc6ef8d74a