Fix testRtcHctosys for changes to hctosys
As of kernel 4.19, hctosys reflects the rtc that set the clock, and
is not set if there is an error. To account for this, we now check the
config for CONFIG_RTC_HCTOSYS_DEVICE, and ensure that if /dev/rtc is
set, it points to the same value as the config, and if not, that the
config is set to rtc0.

Change-Id: I0492d39bb2c94bd2324664ebb52f3fca799cd070
Bug: 158051176
Test: atest KernelApiSysfsTest#testRtcHctosys