bpf_native_test - improve error logging
Seen on dashboards for android-mainline kernel:
  test/vts-testcase/kernel/api/bpf_native_test/BpfTest.cpp:110: Failure
  test/vts-testcase/kernel/api/bpf_native_test/BpfTest.cpp:110: Failure
  Expected: (-1) != (setsockopt(recvSock, 1, 50, &prog_fd, sizeof(prog_fd))), actual: -1 vs -1
  attach bpf program failedBad file descriptor
  ...(repeated)...

which appears to be because we fail to even check that retrieving
the prog_fd (much earlier on) succeeded.

This won't fix the core failure, but hopefully we'll get more information
about what/why it's failing, instead of being overwhelmed by the later spam.

Test: atest bpf_module_test
Signed-off-by: Maciej Żenczykowski <maze@google.com>
Change-Id: I314fef32eb624a877c17a5bcb8810445ce3cfe5d