[bpf_native_test] Remove unused libprocessgroup dependency
Bug: http://b/197965342

libutils.so, a dependency of bpf_native_test, depends on
libprocessgroup.so.  If we include libprocessgroup.a for bpf_native_test
as well, the upstream lld linker tries to unnecessarily fetch symbols
from libprocessgroup.a to redefine undefined symbols from libutils.so.
Remove the dependence since it seems unused anyway.

Test: build with aosp/1809741 which has lld with the above behavior.
Change-Id: I49a453ccc23831fe02a97411cbd8f78edeb13a61