vts_kernel_encryption_test: support decryption
So far the cipher implementations in vts_kernel_encryption_test have
only needed to support encryption, since the tests compute the expected
ciphertext rather than decrypting the actual ciphertext.

That won't work for the f2fs compression+encryption test, because
different implementations of the same compression algorithm can produce
different outputs.  So we'll need to do the test in the other way.

That requires decryption support, so add it.

Bug: 162253911
Test: 'atest -v vts_kernel_encryption_test' on Cuttlefish
Change-Id: I994cb9ec385424da31bbd6b0a15d9940dd651c1d