vts_kernel_encryption_test: test f2fs compression+encryption
f2fs recently added support for per-file compression.  Compression can
be used in combination with encryption, in which case the compressed
data is encrypted rather than the uncompressed data.

We need to verify that the encryption is actually being done as expected
in this case.  So add a test which verifies it.

This is adapted from a similar test which I wrote for xfstests
(https://git.kernel.org/pub/scm/fs/xfs/xfstests-dev.git/commit/?id=8b28dae4ac8c0cc614033df04eef2b5ce21b4458).

Bug: 162253911
Test: 'atest -v vts_kernel_encryption_test' on Cuttlefish
Change-Id: I3c76f259645bed5f29d4f99d45651e13a46855a5